requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
entrypoints==0.3
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
googleapis-common-protos==1.6.0
grpcio==1.25.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.1.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.3
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.0
nbconvert==5.6.1
nbformat==4.4.0
neuralgym==0.0.1
notebook==6.0.2
numpy==1.17.4
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==3.0.2
protobuf==3.11.0
ptyprocess==0.6.0
py==1.8.0
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.6
pytest==5.3.1
python-dateutil==2.8.1
PyYAML==5.2
pyzmq==18.1.1
requests==2.22.0
Send2Trash==1.5.0
six==1.13.0
tensorboard==1.15.0
tensorflow-datasets==1.3.0
tensorflow-estimator==1.15.1
tensorflow-metadata==0.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.40.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
